backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.4.0
openai==1.3.7
optuna==3.4.0
# claude-ai is not a standard pypi package, assuming intent or using openai compatible for now
# or keeping as requested if available, but usually anthropic is the package
anthropic
python-dotenv==1.0.0
